smic-pytorch/requirements.txt
matplotlib==3.7.3
minepy==1.2.6
numpy==1.24.4
pandas==2.0.3
Pillow==10.0.0
POT==0.9.1
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.10.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
